+++ lib/glimmer/libui/checkbox_text_column_proxy.rb
@@ -1,0 +1,44 @@
+# Copyright (c) 2021 Andy Maleh
+#
+# Permission is hereby granted, free of charge, to any person obtaining
+# a copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including
+# without limitation the rights to use, copy, modify, merge, publish,
+# distribute, sublicense, and/or sell copies of the Software, and to
+# permit persons to whom the Software is furnished to do so, subject to
+# the following conditions:
+#
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+# LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+require 'glimmer/libui/control_proxy'
+require 'glimmer/libui/column'
+require 'glimmer/libui/dual_column'
+require 'glimmer/libui/editable_column'
+
+module Glimmer
+  module LibUI
+    # Proxy for LibUI checkbox text column objects
+    #
+    # Follows the Proxy Design Pattern
+    class CheckboxTextColumnProxy < ControlProxy
+      include Column
+      include DualColumn
+      include EditableColumn # TODO split into editable_checkbox and editable_text properties in the future given that checkbox and text can be edited separately
+      
+      private
+      
+      def build_control
+        @parent_proxy.append_checkbox_text_column(name, column_index, -1, second_column_index, editable_value)
+      end
+    end
+  end
+end